requirements.txt
pyspark==3.4.1
psycopg2-binary==2.9.9
pandas==2.1.1
